requirements.txt
aiohttp==3.8.3
aiohttp-graphql==1.1.0
aiosignal==1.3.1
aniso8601==7.0.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
charset-normalizer==2.1.1
frozenlist==1.3.3
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==2.0.1
idna==3.4
multidict==6.0.4
promise==2.3
psycopg2==2.9.5
python-dotenv==0.21.0
Rx==1.6.3
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.45
yarl==1.8.2
